Index.cshtml inline css update
This is not perfectly fixed. However it reshuffles the page around to read left to right, from our previous elements being a bit unordered to follow. 
Furthermore, we need more information supplied into the framework for our audiences to understand and be captivated by our product, this is a change that needs to be worked on.